pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coreops"
version = "0.1.0"
description = "CoreOps backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "celery>=5.3",
  "redis>=5.0",
  "httpx>=0.27.0",
  "jinja2>=3.1",
  "python-multipart>=0.0.9"
]

[tool.setuptools]
packages = ["app"]